refactor: extract media library asset helpers
