+++ src/features/media-library/services/media-library-service.ts
@@ -26,190 +26,18 @@ import {
 import { gifFrameCache } from '@/features/media-library/deps/timeline-services';
 import { opfsService } from './opfs-service';
 import { proxyService } from './proxy-service';
+import { ensureFileHandlePermission, FileAccessError } from './file-access';
+import {
+  buildGeneratedMediaOpfsPath,
+  getGeneratedImageDimensions,
+  getThumbnailDimensions,
+  persistGeneratedMediaAsset,
+} from './media-asset-helpers';
 import { validateMediaFile, getMimeType } from '../utils/validation';
 import { getSharedProxyKey } from '../utils/proxy-key';
 import { mediaProcessorService } from './media-processor-service';
 import { generateThumbnail } from '../utils/thumbnail-generator';
-
-function getImageDimensionsFromBitmap(bitmap: ImageBitmap): { width: number; height: number } {
-  return {
-    width: bitmap.width,
-    height: bitmap.height,
-  };
-}
-
-async function getGeneratedImageDimensions(file: File): Promise<{ width: number; height: number }> {
-  if (typeof createImageBitmap === 'function') {
-    const bitmap = await createImageBitmap(file);
-    const dimensions = getImageDimensionsFromBitmap(bitmap);
-    bitmap.close();
-    return dimensions;
-  }
-
-  return new Promise((resolve, reject) => {
-    const image = new Image();
-    const url = URL.createObjectURL(file);
-
-    image.onload = () => {
-      URL.revokeObjectURL(url);
-      resolve({
-        width: image.naturalWidth,
-        height: image.naturalHeight,
-      });
-    };
-
-    image.onerror = () => {
-      URL.revokeObjectURL(url);
-      reject(new Error('Failed to load generated image'));
-    };
-
-    image.src = url;
-  });
-}
-
-function getThumbnailDimensions(
-  width: number,
-  height: number,
-  maxSize: number
-): { width: number; height: number } {
-  const safeWidth = Math.max(1, Math.round(width));
-  const safeHeight = Math.max(1, Math.round(height));
-
-  if (safeWidth >= safeHeight) {
-    return {
-      width: maxSize,
-      height: Math.max(1, Math.floor(maxSize * safeHeight / safeWidth)),
-    };
-  }
-
-  return {
-    width: Math.max(1, Math.floor(maxSize * safeWidth / safeHeight)),
-    height: maxSize,
-  };
-}
-
-async function blobToArrayBuffer(blob: Blob): Promise<ArrayBuffer> {
-  if (typeof blob.arrayBuffer === 'function') {
-    return blob.arrayBuffer();
-  }
-
-  return new Response(blob).arrayBuffer();
-}
-
-function buildGeneratedMediaOpfsPath(mediaId: string): string {
-  return `content/${mediaId.slice(0, 2)}/${mediaId.slice(2, 4)}/${mediaId}/data`;
-}
-
-interface PersistGeneratedMediaOptions {
-  file: File;
-  projectId: string;
-  mediaMetadata: MediaMetadata;
-  thumbnailBlob?: Blob;
-  thumbnailWidth?: number;
-  thumbnailHeight?: number;
-}
-
-async function persistGeneratedMediaAsset({
-  file,
-  projectId,
-  mediaMetadata,
-  thumbnailBlob,
-  thumbnailWidth,
-  thumbnailHeight,
-}: PersistGeneratedMediaOptions): Promise<MediaMetadata> {
-  let metadataCreated = false;
-  let thumbnailSaved = false;
-
-  try {
-    if (!mediaMetadata.opfsPath) {
-      throw new Error('Generated media is missing an OPFS path.');
-    }
-
-    await opfsService.saveFile(mediaMetadata.opfsPath, await blobToArrayBuffer(file));
-
-    if (thumbnailBlob && thumbnailWidth && thumbnailHeight) {
-      const thumbnailId = crypto.randomUUID();
-
-      await saveThumbnailDB({
-        id: thumbnailId,
-        mediaId: mediaMetadata.id,
-        blob: thumbnailBlob,
-        timestamp: 0,
-        width: thumbnailWidth,
-        height: thumbnailHeight,
-      });
-
-      mediaMetadata.thumbnailId = thumbnailId;
-      thumbnailSaved = true;
-    }
-
-    await createMediaDB(mediaMetadata);
-    metadataCreated = true;
-    await associateMediaWithProject(projectId, mediaMetadata.id);
-    return mediaMetadata;
-  } catch (error) {
-    if (metadataCreated) {
-      try {
-        await deleteMediaDB(mediaMetadata.id);
-      } catch (cleanupError) {
-        logger.warn(`Failed to roll back generated media metadata ${mediaMetadata.id}:`, cleanupError);
-      }
-    }
-
-    if (thumbnailSaved) {
-      try {
-        await deleteThumbnailsByMediaId(mediaMetadata.id);
-      } catch (cleanupError) {
-        logger.warn(`Failed to roll back generated thumbnail ${mediaMetadata.id}:`, cleanupError);
-      }
-    }
-
-    if (mediaMetadata.opfsPath) {
-      try {
-        await opfsService.deleteFile(mediaMetadata.opfsPath);
-      } catch (cleanupError) {
-        logger.warn(`Failed to roll back generated OPFS file ${mediaMetadata.opfsPath}:`, cleanupError);
-      }
-    }
-
-    throw error;
-  }
-}
-/**
- * Check and request permission for a file handle
- * @returns true if permission granted, false otherwise
- */
-async function ensureFileHandlePermission(
-  handle: FileSystemFileHandle
-): Promise<boolean> {
-  try {
-    // Check current permission state
-    const permission = await handle.queryPermission({ mode: 'read' });
-    if (permission === 'granted') {
-      return true;
-    }
-
-    // Request permission if not granted
-    const newPermission = await handle.requestPermission({ mode: 'read' });
-    return newPermission === 'granted';
-  } catch (error) {
-    logger.error('Failed to get file handle permission:', error);
-    return false;
-  }
-}
-
-/**
- * Error thrown when file handle permission is denied or file is missing
- */
-export class FileAccessError extends Error {
-  constructor(
-    message: string,
-    public readonly type: 'permission_denied' | 'file_missing' | 'unknown'
-  ) {
-    super(message);
-    this.name = 'FileAccessError';
-  }
-}
+export { FileAccessError } from './file-access';
 
 /**
  * Media Library Service - Coordinates OPFS + IndexedDB + metadata extraction
